Add player info output
For user login methods, add tests on the role and get player info if user is in player role
