+++ src/TestClients/IdentityServerTestClient/CommandBase.cs
@@ -3,6 +3,11 @@ using Newtonsoft.Json.Linq;
 using System;
 using System.Net.Http;
 using System.Threading.Tasks;
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
+using IdentityModel;
+using System.Net.Http.Headers;
+using System.Net;
 
 namespace IdentityServerTestClient
 {
@@ -26,7 +31,7 @@ namespace IdentityServerTestClient
             {
                 return await ExecuteAsync();
             }
-            catch(CommandArgumentException cae)
+            catch (CommandArgumentException cae)
             {
                 Console.WriteLine(cae.Message);
                 return -1;
@@ -54,6 +59,47 @@ namespace IdentityServerTestClient
 
             var content = await response.Content.ReadAsStringAsync();
             Console.WriteLine(JObject.Parse(content));
+        }
+        protected async Task ShowPlayerInfoAsync(string accessToken)
+        {
+            var role = GetClaimFromToken(accessToken, JwtClaimTypes.Role);
+            Console.WriteLine($"Role in token is '{role}'");
+            if (role == "Player")
+            {
+                var player = await GetPlayerInfoAsync(accessToken);
+                Console.WriteLine(player.ToString());
+            }
+        }
+
+        protected async Task<dynamic> GetPlayerInfoAsync(string accessToken)
+        {
+            // call api
+            var client = new HttpClient();
+            client.SetBearerToken(accessToken);
+            client.DefaultRequestHeaders.Accept.Clear();
+            client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+
+            var response = await client.GetAsync("http://localhost:5000/api/player");
+            if (response.StatusCode == HttpStatusCode.NotFound)
+            {
+                return null;
+            }
+            if (!response.IsSuccessStatusCode)
+            {
+                Console.WriteLine(response.StatusCode);
+            }
+
+            var content = response.Content.ReadAsStringAsync().Result;
+            dynamic result = JToken.Parse(content);
+            return result.player;
+        }
+
+        protected string GetClaimFromToken(string accessToken, string claimType)
+        {
+            return new JwtSecurityToken(accessToken)
+                .Claims
+                .FirstOrDefault(c => c.Type == claimType)
+                ?.Value;
         }
     }
 }
